Strict types and move Exceptions to dedicated namespace
