+++ composer.json
@@ -17,7 +17,7 @@
         }
     ],
     "require": {
-      "php": "^7.3|^8.0"
+      "php": "^7.4|^8.0"
     },
     "scripts": {
       "lint": "find . -name \\*.php -not -path './vendor/*' -print0 | xargs -0 -n1 php -l",
