+++ lib/Cron/Cleanup.php
@@ -29,7 +29,7 @@ declare(strict_types=1);
 namespace OCA\Text\Cron;
 
 use OCA\Text\Db\Session;
-use OCA\Text\DocumentHasUnsavedChangesException;
+use OCA\Text\Exception\DocumentHasUnsavedChangesException;
 use OCA\Text\Service\DocumentService;
 use OCA\Text\Service\ImageService;
 use OCA\Text\Service\SessionService;
