+++ lib/Exception/DocumentHasUnsavedChangesException.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2019 Julius Härtl <jus@bitgrid.net>
  *
@@ -21,7 +24,7 @@
  *
  */
 
-namespace OCA\Text;
+namespace OCA\Text\Exception;
 
 class DocumentHasUnsavedChangesException extends \Exception {
 }
