+++ lib/Exception/VersionMismatchException.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2019 Julius Härtl <jus@bitgrid.net>
  *
@@ -21,7 +24,12 @@
  *
  */
 
-namespace OCA\Text;
+namespace OCA\Text\Exception;
 
-class DocumentSaveConflictException extends \Exception {
+use OCP\AppFramework\Http;
+
+class VersionMismatchException extends \Exception {
+	public function getStatus(): int {
+		return Http::STATUS_PRECONDITION_FAILED;
+	}
 }
