+++ lib/Service/ApiService.php
@@ -29,10 +29,10 @@ namespace OCA\Text\Service;
 use Exception;
 use OC\Files\Node\File;
 use OCA\Text\AppInfo\Application;
-use OCA\Text\DocumentHasUnsavedChangesException;
-use OCA\Text\DocumentSaveConflictException;
+use OCA\Text\Exception\DocumentHasUnsavedChangesException;
+use OCA\Text\Exception\DocumentSaveConflictException;
 use OCA\Text\TextFile;
-use OCA\Text\VersionMismatchException;
+use OCA\Text\Exception\VersionMismatchException;
 use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
