+++ lib/Service/DocumentService.php
@@ -45,9 +45,9 @@ use OCA\Text\Db\Document;
 use OCA\Text\Db\DocumentMapper;
 use OCA\Text\Db\Step;
 use OCA\Text\Db\StepMapper;
-use OCA\Text\DocumentHasUnsavedChangesException;
-use OCA\Text\DocumentSaveConflictException;
-use OCA\Text\VersionMismatchException;
+use OCA\Text\Exception\DocumentHasUnsavedChangesException;
+use OCA\Text\Exception\DocumentSaveConflictException;
+use OCA\Text\Exception\VersionMismatchException;
 use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Db\Entity;
 use OCP\Constants;
